requirements.txt
langchain
python-dotenv
langchain_community
sympy
langgraph
azure-search-documents
azure-identity
chainlit



